\name{baba}
\alias{baba}
\docType{data}
\title{Nishisato and Baba, 1999, artificial set of data}
\description{
18 subjects and 5 items with different numbers of options.
}
\usage{data(baba)}
\format{
  A data frame with 18 subjects on 5 items with the following numbers of options:
  \describe{
    \item{\code{V1}}{2 options}
    \item{\code{V2}}{3 options}
    \item{\code{V3}}{4 options}
    \item{\code{V4}}{5 options}
    \item{\code{V5}}{6 options}
  }
}
\source{
\describe{
\item{}{
Nishisato S, Baba Y (1999). On contingency, projection and forced classification of dual scaling. \emph{Behaviormetrika}, 26, 207--219.}
}
}

\references{
	\describe{
 \item{}{Nishisato S (2007). \emph{Multidimensional Nonlinear Descriptive Analysis}. Chapman & Hall/CRC.}
}
}
\examples{
data(baba)
}
\keyword{datasets}
