% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bad_coded.R
\docType{data}
\name{bad_coded}
\alias{bad_coded}
\title{Nishisato and Clavel, artificial set of data}
\format{
A data set with 10 observations on the following 3 variables:
\describe{
\item{\code{V2}}{Option 1 is omited}
\item{\code{V3}}{Options go from 1 to 8}
\item{\code{V4}}{Option 1 omited, other are not consecutive and there is NA}
}
}
\source{
Nishisato, S. and Baba, Y. (1999). On contingency, projection and
forced classification of dual scaling. \emph{Behaviormetrika}, 26, 207--219.
}
\usage{
bad_coded
}
\description{
10 observation and 3 variables erroneously coded.
}
\details{
The data were collected from 23 participants at a workshop in
Singapore in 1985
}
\references{
Nishisato, S. (2007). \emph{Multidimensional Nonlinear
Descriptive Analysis}. Chapman & Hall/CRC.
}
\keyword{datasets}
