% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/christmas.R
\docType{data}
\name{christmas}
\alias{christmas}
\title{Christmas party plans}
\format{
A subset of the original data set of 14 subjects on 15 pairs of
6 plans:
\describe{
\item{\code{plan1}}{A pub/restaurant crawl after work}
\item{\code{plan2}}{A reasonably priced lunch in an area restaurant}
\item{\code{plan3}}{Keep to one's self}
\item{\code{plan4}}{An evening banquet at a restaurant}
\item{\code{plan5}}{A pot-luck at someone's home after work}
\item{\code{plan6}}{A ritzy lunch at a good restaurant (tablecloths)}
}
}
\source{
Nishisato, S. and Nishisato, I.(1994). \emph{Dual Scaling in a
Nutshell}. Toronto: MicroStats.
}
\usage{
christmas
}
\description{
As a course assignment for Nishisato's class, Ian Wiggins, a
student, collected paired comparison data from 14 researchers at a research
institute on his eight Christmas party plans.
}
\details{
The data were originally collected from 14 participants by 28 pairs
of plans with elements 1 for the choice of the first plan and 2 for the
choice of the second plan. For computations, \code{2} is converted to \code{-1}.
}
\references{
Nishisato, S. (2022). Optimal Quantification and Symmetry.
\emph{Behaviormetrika}, 12, 137.
}
\keyword{datasets}
