% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curricula.R
\docType{data}
\name{curricula}
\alias{curricula}
\title{Curricula and Social classes}
\format{
A data set of 390 subjects on 5 social classes and 3 different
curricula:
\describe{
\item{\code{s.class1}}{Merged social classes I and II because the
frequencies were small}
\item{\code{s.class2}}{Social class III}
\item{\code{s.class3}}{Social class IV}
\item{\code{s.class4}}{Social class V}
\item{\code{curricula1}}{College Prep curriculum}
\item{\code{curricula2}}{General curriculum}
\item{\code{curricula3}}{Commercial curriculum}
}
}
\source{
Nishisato, S. and Nishisato, I.(1994). \emph{Dual Scaling in a
Nutshell}. Toronto: MicroStats.
}
\usage{
curricula
}
\description{
Hollingshead (1949) found that the members of a small Middle
Western community in the United States divided themselves into 5 social
classes. He investigated his prediction that adolescents in the different
social classes would enroll in different curricula
}
\details{
The data were originally collected from 390 participants in terms
of their social classes and actual curriculum enrollments.
}
\references{
Hollingshead, A.B. (1949). Elmtown's Youth: The Impact of
Social Classes on Adolescents. \emph{Wiley}.
}
\keyword{datasets}
