% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_cf.R
\name{ds_cf}
\alias{ds_cf}
\title{Contingency and frequency data analysis}
\usage{
ds_cf(input, solutions = NULL)
}
\arguments{
\item{input}{A data set with valid data}

\item{solutions}{Optional arguments. A number of intended solutions}
}
\value{
\item{call}{Call with all of the specified arguments are specified by
their full names}
\item{orig_data}{Initial data}
\item{item_op_lbl}{Item options labels}
\item{sub_lbl}{Subjects options labels}
\item{tot_row}{Sum of subject values}
\item{tot_row}{Sum of item values}
\item{solutions}{Maximum possible solutions}
\item{out}{Results obtained}
\item{norm_opt}{Normed option weights}
\item{proj_opt}{Projected option weights}
\item{norm_sub}{Normed subject scores}
\item{proj_sub}{Projected subject scores}
\item{appro0}{Order 0 approximation for initial data}
\item{approx}{Order \code{k} approximation for each solution}
\item{residual0}{Residual matrix for initial data}
\item{residual}{Residual matrix \code{k} for each solution}
}
\description{
Contingency and frequency data analysis
}
\examples{
ds_cf(curricula)
ds_cf(preferences)
}
