% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ds_eigen}
\alias{ds_eigen}
\title{Calculate SVD decomposition of the matrix}
\usage{
ds_eigen(obj)
}
\arguments{
\item{obj}{A Matrix}
}
\value{
Result of \code{svd} function
}
\description{
Calculate SVD decomposition of the matrix
}
\keyword{internal}
