% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_mc_check.R
\name{ds_mc_check}
\alias{ds_mc_check}
\title{Function to identify incorrect Multiple Choice input data}
\usage{
ds_mc_check(input, mode = c("rad", "act"))
}
\arguments{
\item{input}{The input data to be checked}

\item{mode}{Do you want to use a radical ("rad") correction mode
or active ("act") allocations?}
}
\value{
A list with the original input and the converted input
}
\description{
Function to identify incorrect Multiple Choice input data
}
\examples{
ds_mc_check(singaporean)
ds_mc_check(bad_coded)
}
