% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ds_out_tbl}
\alias{ds_out_tbl}
\title{Generate the dataframe out}
\usage{
ds_out_tbl(inf, eig, op)
}
\arguments{
\item{inf}{List with required row size information}

\item{eig}{List with the information of the values to be displayed}

\item{op}{Function where each output is used: (1) ds_cf (2) ds_mc (3)
ds_mc_A (4) ds_mc_C (5) ds_pc / ds_ro}
}
\value{
A dataframe with the eigenvalue's information of the process
}
\description{
Generate the dataframe out
}
\keyword{internal}
