% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_pc.R
\name{ds_pc}
\alias{ds_pc}
\title{Paired comparison data analysis}
\usage{
ds_pc(input, solutions = NULL)
}
\arguments{
\item{input}{A data set with valid data}

\item{solutions}{Optional argument. A number of intended solutions}
}
\value{
\item{call}{Call with all of the specified arguments are specified by
their full names}
\item{orig_data}{Initial data}
\item{item_op_lbl}{Item options labels}
\item{sub_lbl}{Subjects options labels}
\item{solutions}{Maximum possible solutions}
\item{out}{Results obtained}
\item{mat_e}{Matrix E}
\item{norm_opt}{Normed option weights}
\item{proj_opt}{Projected option weights}
\item{norm_sub}{Normed subject scores}
\item{proj_sub}{Projected subject scores}
}
\description{
Paired comparison data analysis
}
\examples{
ds_pc(christmas)
}
