% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_ro.R
\name{ds_ro}
\alias{ds_ro}
\title{Rank order data analysis}
\usage{
ds_ro(input, solutions = NULL)
}
\arguments{
\item{input}{A data set with valid data}

\item{solutions}{Optional argument. A number of intended solutions}
}
\value{
\item{call}{Call with all of the specified arguments are specified by
their full names}
\item{orig_data}{Initial data}
\item{item_op_lbl}{Item options labels}
\item{sub_lbl}{Subjects options labels}
\item{solutions}{Maximum possible solutions}
\item{out}{Results obtained}
\item{mat_e}{Matrix E}
\item{norm_opt}{Normed option weights}
\item{proj_opt}{Projected option weights}
\item{norm_sub}{Normed subject scores}
\item{proj_sub}{Projected subject scores}
\item{out_rank}{Results obtained by rank analysis}
\item{norm_opt_rank}{Normed option weights by rank analysis}
\item{proj_opt_rank}{Projected option weights by rank analysis}
\item{norm_rank}{Normed rank scores}
\item{proj_rank}{Projected rank scores}
}
\description{
Rank order data analysis
}
\examples{
ds_ro(goverment)
}
