% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ds_select_solutions}
\alias{ds_select_solutions}
\title{Selection of solutions for analysis}
\usage{
ds_select_solutions(sols, max)
}
\arguments{
\item{sols}{The number of solutions requested}

\item{max}{Maximum possible data solutions}
}
\value{
Solutions used in further analysis
}
\description{
Selection of solutions for analysis
}
\keyword{internal}
