% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferences.R
\docType{data}
\name{preferences}
\alias{preferences}
\title{Preferences, artificial set of data}
\format{
A data set of contingency responses:
\describe{
\item{\code{A, B, C}}{Do you prefer coffee to tea? (Yes, Not always, No)}
\item{\code{Y, N}}{Do you smoke? (Yes, No)}
}
}
\source{
Nishisato, S. (1980). \emph{Analysis of categorical data: Dual
Scaling and its Applications}. University of Toronto: Heritage.
}
\usage{
preferences
}
\description{
Artificial set of data where 13 people were asked two questions.
}
\details{
Artificial set of data where 13 people were asked two questions.
}
\references{
Nishisato, S. (1980). \emph{Analysis of categorical data: Dual
Scaling and its Applications}. University of Toronto: Heritage.
}
\keyword{datasets}
