% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.dualScale}
\alias{print.dualScale}
\title{Print of Dual Scale analysis}
\usage{
\method{print}{dualScale}(x, ...)
}
\arguments{
\item{x}{An dualScale object for which a summary is desired}

\item{...}{Arguments to be passed to methods}
}
\value{
A print of the available information from the object
}
\description{
\code{print} method for package "dualScale"
}
\examples{
print(ds_cf(curricula))
print(ds_cf(preferences))
print(ds_mc(singaporean))
print(ds_mcf(singaporean, crit = 1))
print(ds_pc(christmas))
print(ds_ro(goverment))
}
\seealso{
\code{\link[=print]{print()}}
}
