\name{singapore}
\alias{singapore}
\docType{data}
\title{
Singaporean children as viewed by adults}
\description{A short survey on childrem in Singapore.
}
\usage{data(singapore)}
\format{
  A data set of 23 subjects on 4 multiple-choice questions.
  \describe{
    \item{\code{V1}}{How old are you? (1 = 20-29, 2 = 30-39, 3 = 40 or over)}
    \item{\code{V2}}{Children today are not as disciplined as when I was a child. (1 = agree, 2 = disagree, 3 = I cannot tell)}
    \item{\code{V3}}{Children today are not as fortunate as when I was a child. (1 = agree, 2 = disagree, 3 = I cannot tell) }
    \item{\code{V4}}{Religions should be taught at school. (1 = agree, 2 = disagree, 3 = Indifferent)}
  }
}
\details{The data were collected from 23 participants at a workshop in Singapore in 1985}
\source{
\describe{	
\item{}{Nishisato S and Nishisato, I.(1994). \emph{Dual Scaling in a Nutshell}. Toronto: MicroStats.}
}
}
\references{
Nishisato S (2007). \emph{Multidimensional Nonlinear Descriptive Analysis}. Chapman & Hall/CRC.}

\examples{
data(singapore)
}
\keyword{datasets}
