% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singaporean.R
\docType{data}
\name{singaporean}
\alias{singaporean}
\title{Singaporean children as viewed by adults}
\format{
A data set of 23 subjects on 4 multiple-choice questions:
\describe{
\item{\code{A}}{How old are you? (1 = 20-29, 2 = 30-39, 3 = 40 or over)}
\item{\code{B}}{Children today are not as disciplined as when I was a
child (1 = agree, 2 = disagree, 3 = I cannot tell)}
\item{\code{C}}{Children today are not as fortunate as when I was a
child (1 = agree, 2 = disagree, 3 = I cannot tell)}
\item{\code{D}}{Religions should be taught at school (1 = agree,
2 = disagree, 3 = Indifferent)}
}
}
\source{
Nishisato, S. and Nishisato, I.(1994). \emph{Dual Scaling in a
Nutshell}. Toronto: MicroStats.
}
\usage{
singaporean
}
\description{
A short survey on childrem in Singapore.
}
\details{
The data were collected from 23 participants at a workshop in
Singapore in 1985
}
\references{
Nishisato, S. (2007). \emph{Multidimensional Nonlinear
Descriptive Analysis}. Chapman & Hall/CRC.
}
\keyword{datasets}
