% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.dualScale}
\alias{summary.dualScale}
\title{Summary of Dual Scale analysis}
\usage{
\method{summary}{dualScale}(object, ...)
}
\arguments{
\item{object}{An dualScale object for which a summary is desired}

\item{...}{Arguments to be passed to methods}
}
\value{
A summary of the available information from the object
}
\description{
\code{summary} method for class "dualScale"
}
\examples{
summary(ds_cf(curricula))
summary(ds_cf(preferences))
summary(ds_mc(singaporean))
summary(ds_mcf(singaporean, crit = 1))
summary(ds_pc(christmas))
summary(ds_ro(goverment))
}
\seealso{
\code{\link[=summary]{summary()}}
}
